+++ R/utils_get_data.R
@@ -46,8 +46,9 @@
   if (mc[1] && rn == colnames(mf)[1]) {
     mc[1] <- FALSE
     if (inherits(x, c("coxph", "flexsurvreg", "coxme", "survreg", "survfit", "crq", "psm"))) {
-      mf <- cbind(mf[[1]][, 1], mf[[1]][, 2], mf)
-      colnames(mf)[1:2] <- rn_not_combined
+      n_of_responses <- ncol(mf[[1]])
+      mf <- cbind(as.data.frame(as.matrix(mf[[1]])), mf)
+      colnames(mf)[1:n_of_responses] <- rn_not_combined
     } else {
       tryCatch(
         {