+++ NEWS.md
@@ -17,6 +17,7 @@ as possible.
 
 * Fixed issue in `get_data()` for `MixMod` objects, which were caused due to internal changes in *GLMMadaptive*.
 * `get_data()` for zero-inflated models from *pscl* did not include the offset-term in cases where the offset was defined as argument, not inside the model formula.
+* Fixed issue in `get_data()` for `coxph` models with survival-objects with `event`-argument as response.
 
 # insight 0.10.0